refactor(lib): extract menu and bootstrap logic into submodules
